+++ src/main/java/com/project/chamjimayo/domain/entity/RestroomManager.java
@@ -22,7 +22,7 @@ public class RestroomManager extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "manager_id")
-	private Integer managerId;
+	private Long managerId;
 
 	// 화장실 아이디 (어느 화장실을 관리하는지)
 	@OneToMany(mappedBy = "restroomManager")
@@ -39,5 +39,4 @@ public class RestroomManager extends BaseEntity {
 	// 이메일
 	@Column(name = "email")
 	private String email;
-
 }
