+++ src/main/java/com/project/chamjimayo/domain/entity/User.java
@@ -25,7 +25,7 @@ public class User extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "user_id")
-	private Integer userId;
+	private Long userId;
 
 	// 이름
 	@Column(name = "name")
@@ -71,7 +71,7 @@ public class User extends BaseEntity {
 
 	@Builder
 	public User(String name, String nickname, Integer point,
-			String gender, Role role, String authId, AuthType authType) {
+		String gender, Role role, String authId, AuthType authType) {
 		this.name = name;
 		this.nickname = nickname;
 		this.point = point;
