+++ src/main/java/com/project/chamjimayo/repository/SearchRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SearchRepository extends JpaRepository<Search, Integer> {
+public interface SearchRepository extends JpaRepository<Search, Long> {
 
 	Optional<Search> findTopByUserAndClickOrderBySearchIdDesc(User user, int i);
